Inlcude "version number" parameter in URLs which loads JS files. Example: http://test.cust.cention.se/Cention.app/Resources/Javascript/errand.js?v=20140711. With this change we should in the future only have to of all customers only tell two that they need to perform a forced reload in their web browsers. Those customers are Nelly and Ongame. The reason is that these customers works 24/7. The version number is the current days date.
The great thing though is that since Nelly does not have this change right now and they
will get it when we update them tomorrow then we shouldn't have to tell them to perform
a forced reload. This means we should hopefully avoid a lot of issues that will be
reported to us tomorrow.