Fix text-value state not updating in Textfield/TextView
TextView and Textfield components are based on the same code which
updates the text-value state in Javascript when the change event
is triggered on the input type text or textfield.

In Firefox I've observed the change event to be triggered if one
of two things happen.
1. If you click outside of the textarea when the caret is in the textarea.
2. When the caret is in the textarea and you change from the browser window to another window.

In IE11 the change event is not triggered for #2.

In CCC there is now a function which allows you to select text in EE
and click a button to copy the text to the errand.
If you for example delete all text in the errand's answer textarea
and then with the caret still in the textarea switches to the EE window
then when you click the button to copy the text the old text that you
removed will be inserted with your selection because the text-value
state hasn't updated in the TextView component.

Now the TextView and Textfield components listens to keyup events
instead of change events and updates the state whenever that event
is triggered.

This however can be to much to handle for web browsers with slower
Javascript engines such as IE7 and IE8.