Fix problem with system asking if you want to send empty answer
This happens even tough you've written an answer.

The problem was that we've made a change to how the text-value state
is set. Previously it was set with the setState() function.
However since text-value is the component's default state
the updateVisual() function is called. This means that the element's value
is reset. This caused in some browsers the caret to be reset to first position.
Since the text-value is updated on keyup it caret was reset on every keyboard click.

The idea was to solve this by setting the text-value directory in the state list.
This then makes sure updateVisual() is not called.
However the first commit did not make the proper change instead the change
that was made caused the system to ask if the user wants to send an empty
answer when the send button was clicked because it thought nothing had been
entered in the text box.