Fix problem with caret being reset to first character on key press
The problem was that on every key press the component's text-value state is updated.
However since text-value is the component's default state
the updateVisual() function is called. This means that the element's value
is reset. This caused in some browsers the caret to be reset to first position.
Since the text-value is updated on keyup it caret was reset on every keyboard click.

Now the text-value state is set directly instead of calling the setState() function.
This then makes sure updateVisual() is not called.