Fix problem with converting string to timestamp
The problem occured if time zone on the server was set to anything
but Europe/Stockholm (CET/CEST).
This was because the code does some hacky things to subtract
the right amount of hours from the timestamp.
The code had hardcoded the number of hours to subtract for time zone
Europe/Stockholm.