fs#4003 - Bugfix "Turn Chat Off/On" was broken by commit 8078472d37
Changes done in commit 8078472d37 (fs#4003 - MCAM.js: use GET for POST with zero data)
will break the "Turn Chat Off/On" ticker when it changes the
XMLHttpRequest method to GET when the "parameters" array is empty. The
method used in MCAM.fireBackgroundEvent must always be POST because it
always send data in addition to the values in the parameters array.