Fix problem causing external expert questions not be sent
The problem was in the Combobox component.
Many customers have large addressbooks. Comboboxes are used
for entering email addresses so that email addresses can be
choosen from the address book aswell as being entered freely.

If the email address or the email address name contained ' characters
Javascript errors would happen on the page. This would cause
the ExternalExpertAddresses component not to be registered on
the query page which caused the Javascript error when the page
tried to get the text value from the component.

Fix the problem by properly escaping the strings before rendering
them as Javascript on a HTML page.