FS#4105: Get MIME type for uploaded file
To fix FS#4105 proper MIME type for files need to be sent when
sending files as attachments in emails.

To be able to do that first it most be possible to get the proper
MIME type. Since all uploaded files goes through this component
make this component fetch the MIME type for the uploaded file.