FS#4226: Add alternative target to Helper component
The Helper component is activated when a target gains focus.
In some places we have two different components in which you
can enter "the same thing" depending on if something is activated or not.

Example: You have two different fields to enter receipt body in area administration
depending on rich text is enabled or not.
The Helper component can now have two targets and activate on both the Textfield component
and the Ckeditor component.