Resolve the problem in same browsers opening cention-app and external app
The problem is when some opening cention-application in a browser it
creates the SSID for the user login credentials. But later if he opened the
external app in same browser different tab then the Cookie(SSID) was sending
via http header and webframwework fetchuser checking validation for SSID. As the request comes from
external app so SSID is valid and user has logged in another tab so its try to logout.

Now it will fix the problem to redirect the external chat to login page.