FS#4669: Rewrite auth middleware to work with old style session handling
Previously in 3.9 and earlier the session handling used cookies
cention-suiteWFUserID and cention-suiteWFPassword to determine if
a user is logged in or not.

This was changed to instead use cention-suiteSSID.
How often the SSID got invalidated caused problems.

The changes to Permission.feh have been reverted and the changes
in this commit makes the Auth middleware work with the old style
cention-suiteWFUserID and cention-suiteWFPassword.