Fix JavaScript error DragDropSelect component when item has ' in text value
The problem was the the JavaScript code being generated would look something like this:
    setSelected('Mathew O'Conner');

This meant that a syntax error was introduced.
I wasn't able to quickly fix this in a pretty way.
What happens now is that the text value is displayed like this:
    Mathew O\\'Conner

I've decided that this is acceptable because we are about to move way completely
from webframework.