Anna raised some errors related to ckeditor which is belongs to `instanceReady`
Problems: 1. Ckeditor setState no more working because the editorIsReady always false.
Resolves: I reverted back the contentDom event instead of 'instanceReady'
The `contentDom` event goes along with `contentDomUnload` which is fired before the content
DOM structure is destroyed. This is the right moment to detach content DOM event listener.
Otherwise browsers like IE or Opera may throw exceptions when accessing elements from the detached document.

contentDom: Event fired when the editor content (its DOM structure) is ready. It is similar to the native DOMContentLoaded
event, but it applies to the editor content. It is also the first event fired after the CKEDITOR.editable is initialized.