CCC-340: Fix drop down showing in wrong place when page is scrolled
The problem was that it wasn't accounted for the number of pixels
the page was scrolled down when it was calculated where the drop down
menu should be displayed when it should be displayed above the button
because there is more space above the button.
Instead the menu was displayed where it should be displayed had the page
been scrolled up fully.