CCC#991 I18N.I(): Do not normalize text before translation
The old text extraction method normalizes the text before storing them
into Master.strings.fe which adds a bit of fuzziness in the stored text.
The function I18N.I() thus normalizes the text as well before locating
it in the translation dictionaries.

Remove this fuzzyness - we want our text database (Master.strings.fe) to
contain exact strings as found in the source codes. Also this follows
the same behavior as gnu-gettext - we never mangle the extracted text.