CS#120 sockwrap.js: ack all inbound event
This lets us know if the front end received the payload. Before this
there was no way to check (from the backend) whether the front end
receives what we sent.

With this change the confirmation on whether the front end receives the
payload can be logged at the backend.

It may seem that we would be calling the callback twice for the case
when the sockwrap.js client already call the callback, but
socket-io.client already have a builtin check that makes the second
callback call a no-op.