CCC-1571 auth: Accept (and trims!) anything that after (and including) @ in username
Note: This will definitely break existing usernames that contains "@" in
their usernames. This commit assumes that no agents have that character
in their usernames in order to simplify the code.

If problem arises due to this commit then we should decide on whether to
change the agent's username or make our code a bit more complicated.