CS-333 auth.Middleware: insert current user into request context too
It was already added in gin context. Which means client code that needs
to get it needs to have a gin context as well. This is not desireable in
sites that don't have gin context, for example in socket io event in go
chatserver code where what we have is only the request object and not
the gin context. The go chatserver works around this by storing the gin
context in a global map associated with the request.

With this small change we can simplify much of the code in
go-chatserver.